Header 99%, falta contagem da pokédex
